+++ src/shogun-core-main/src/main/java/de/terrestris/shoguncore/util/interceptor/secure/WmsRequestInterceptor.java
@@ -4,7 +4,6 @@ import de.terrestris.shoguncore.dao.LayerDao;
 import de.terrestris.shoguncore.model.layer.Layer;
 import de.terrestris.shoguncore.model.layer.source.ImageWmsLayerDataSource;
 import de.terrestris.shoguncore.service.LayerService;
-import de.terrestris.shoguncore.util.application.ShogunCoreContextUtil;
 import de.terrestris.shoguncore.util.interceptor.GeoserverAuthHeaderRequest;
 import de.terrestris.shoguncore.util.interceptor.MutableHttpServletRequest;
 import de.terrestris.shoguncore.util.interceptor.WmsRequestInterceptorInterface;
@@ -12,11 +11,7 @@ import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
-
 import static org.apache.logging.log4j.LogManager.getLogger;
-
-import java.net.URI;
-import java.net.URISyntaxException;
 import java.util.List;
 
 /**
@@ -27,12 +22,7 @@ import java.util.List;
  * @author Johannes Weskamm
  *
  */
-public class WmsRequestInterceptor implements WmsRequestInterceptorInterface {
-
-    /**
-     *
-     */
-    private URI appUri = null;
+public class WmsRequestInterceptor extends BaseInterceptor implements WmsRequestInterceptorInterface {
 
     /**
      *
@@ -138,6 +128,7 @@ public class WmsRequestInterceptor implements WmsRequestInterceptorInterface {
                     source.getUrl().equalsIgnoreCase(request.getContextPath() +
                     "/geoserver.action")) {
                     match = true;
+                    break;
                 }
             }
         }
@@ -152,34 +143,6 @@ public class WmsRequestInterceptor implements WmsRequestInterceptorInterface {
      */
     private boolean isAllowed(MutableHttpServletRequest request) {
         return isAllowed(request, "layers");
-    }
-
-    /**
-     * <p>forbidRequest.</p>
-     *
-     * @param request a {@link de.terrestris.shoguncore.util.interceptor.MutableHttpServletRequest} object.
-     * @return a {@link de.terrestris.shoguncore.util.interceptor.MutableHttpServletRequest} object.
-     */
-    protected MutableHttpServletRequest forbidRequest(MutableHttpServletRequest request) {
-        this.setAppUriFromRequest(request);
-        String redirectUri = appUri == null ? "" : appUri.toString();
-        request.setRequestURI(redirectUri + "/response/forbidden.action");
-        return request;
-    }
-
-    /**
-     * <p>setAppUriFromRequest.</p>
-     *
-     * @param request a {@link de.terrestris.shoguncore.util.interceptor.MutableHttpServletRequest} object.
-     */
-    protected void setAppUriFromRequest(MutableHttpServletRequest request) {
-        if (appUri == null) {
-            try {
-                appUri = ShogunCoreContextUtil.getApplicationURIFromRequest(request);
-            } catch (URISyntaxException e) {
-                // pass
-            }
-        }
     }
 
 }
